refactor(core): address code-review findings on ADR-0002
Three fixes from the /code-review (Standards + Spec axes):

1. (Standards, hard) Tighten auth::app_ticket visibility to pub(crate).
   After apply_app_ticket was deleted, pub mod app_ticket had no pub
   items — a public path to nothing. Now pub(crate), matching the
   minimal-public-API rule (core AGENTS.md). CHANGELOG breaking entry
   extended to note the mod-path tightening (zero practical impact:
   no pub items remained).

2. (Spec) ADR-0002 decision-table #1 over-stated 'tracing' for the
   resend bootstrap. UnifiedRequestBuilder::build provides header/
   timeout (not tracing spans — those live in Transport, which the
   bootstrap bypass skips). Corrected to 'header/timeout'; the
   tracing gain is the forward-looking rationale (future tracing/retry
   on the request path auto-covers resend), now stated accurately.

3. (Spec) Honor ADR migration-stage-1 'tests move with code': relocate
   the 26 determine_token_type / validate_token_type tests from
   http::test into auth/policy::tests (verbatim; field_reassign_with_default
   allow carried over to match http's pattern). validate() + content_disposition
   tests stay in http. Coverage unchanged (407 lib tests before and after).

Verification: 407 lib + 32 contract tests green; fmt clean; clippy --all-features
and --no-default-features both clean (-Dwarnings).
