+++ crates/openlark-core/src/auth/mod.rs
@@ -1,10 +1,10 @@
 //! 鉴权模块（ADR-0002）：收口「给请求做鉴权」整条 concern——
 //! 获取（acquisition）+ 决策/授权校验（policy）+ 恢复（app_ticket）。
 //!
-//! `acquisition` / `policy` 为 `pub(crate)`：仅 core 内部可达，不进公开 API。
+//! `acquisition` / `policy` / `app_ticket` 为 `pub(crate)`：仅 core 内部可达，不进公开 API。
 
 pub(crate) mod acquisition;
-pub mod app_ticket;
+pub(crate) mod app_ticket;
 pub(crate) mod policy;
 pub mod token_provider;
